Test all rngs against a C reference implementation
