+++ prng/src/bench.ml
@@ -5,7 +5,6 @@ module Command = Core.Command
 module Random_Stdlib       = MakeRandom.Basic(Random)
 module Random_X256pp_Int   = MakeRandom.Basic(X256pp.Int)
 module Random_X256pp_Int64 = MakeRandom.Basic(X256pp.Int64)
-module Random_X256pp_C     = MakeRandom.Basic(X256pp.C)
 
 let bits () =
   Command.run @@ B.make_command [
@@ -17,10 +16,6 @@ let bits () =
         X256pp.Int64.bits ());
     B.Test.create ~name:"X256pp.Int64.bits62" (fun () ->
         X256pp.Int64.bits62 ());
-    B.Test.create ~name:"X256pp.C.bits" (fun () ->
-        X256pp.C.bits ());
-    B.Test.create ~name:"X256pp.C.bits62" (fun () ->
-        X256pp.C.bits62 ());
   ]
 
 let float () =
@@ -29,14 +24,10 @@ let float () =
         Random.float 1.);
     B.Test.create ~name:"X256pp.Int64.float" (fun () ->
         X256pp.Int64.float 1.);
-    B.Test.create ~name:"X256pp.C.float" (fun () ->
-        X256pp.C.float 1.);
     B.Test.create ~name:"MakeRandom(X256pp.Int).float" (fun () ->
         Random_X256pp_Int.float 1.);
     B.Test.create ~name:"MakeRandom(X256pp.Int64).float" (fun () ->
         Random_X256pp_Int64.float 1.);
-    B.Test.create ~name:"MakeRandom(X256pp.C).float" (fun () ->
-        Random_X256pp_C.float 1.);
   ]
 
 let int64 () =
@@ -47,8 +38,6 @@ let int64 () =
         Random_X256pp_Int.int64 Int64.max_int);
     B.Test.create ~name:"MakeRandom(X256pp.Int64).int64" (fun () ->
         Random_X256pp_Int64.int64 Int64.max_int);
-    B.Test.create ~name:"MakeRandom(X256pp.C).int64" (fun () ->
-        Random_X256pp_C.int64 Int64.max_int);
   ]
 
 let () =
