--- prng/src/splitmix/reference.ml
@@ -1,2 +1,0 @@
-external seed: int64 -> unit = "caml_sm64_seed"
-external next: unit -> int64 = "caml_sm64_next"
