+++ prng/src/xoshiro/dune
@@ -1,10 +1,3 @@
 (library
  (name xoshiro)
- (foreign_archives xoshiro)
- (foreign_stubs (language c) (names stubs)))
-
-(foreign_library
- (archive_name xoshiro)
- (language c)
- (flags -lsplitmix)
- (names reference))
+ (libraries splitmix))
