--- prng/src/xoshiro/reference.ml
@@ -1,2 +1,0 @@
-external seed: int64 -> unit = "caml_x256pp_seed"
-external next: unit -> int64 = "caml_x256pp_next"
