+++ prng/src/xoshiro/tests/int.ml
@@ -10,11 +10,6 @@ let int64_to_int_int n =
   let lo = to_int (logand n 0xFFFFFFFFL) in
   (hi, lo)
 
-let test_compare nb f g =
-  for _ = 1 to nb do
-    assert (f () = g ())
-  done
-
 let ocaml_seed seed =
   let s = ref seed in
   for i = 0 to 3 do
@@ -24,13 +19,6 @@ let ocaml_seed seed =
   done
 
 let () =
-  (* Chosen by fair dice rool. Guaranted to be random. *)
-  (* https://xkcd.com/221/ *)
-  let seed = 0xc6bb1e08d1a33f2fL in
-  Reference.seed seed;
-  ocaml_seed seed;
-  test_compare
-    1_000_000
-    Reference.next
-    (fun () ->
-      X256pp.Int.next () |> int_int_to_int64)
+  Testutils.run
+    ~seed_f:ocaml_seed
+    ~next:(fun () -> X256pp.Int.next () |> int_int_to_int64)
