+++ prng/src/xoshiro/tests/x256pp_c.c
@@ -2,5 +2,5 @@
 #include "../../testutils/testutils.h"
 
 int main(int argc, char* argv[]) {
-  return testutils_run(argc, argv, sm64_seed, sm64_next);
+  return testutils_run(argc, argv, x256pp_seed, x256pp_next);
 }
